\docType{class}
\name{Constant}
\alias{Constant}
\alias{Constant-class}
\title{The CONSTANT class}
\description{
Use this class for keeping constants
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{path2scidbBin}:}{Object of class
  \code{"character"}, Path to the bin folder in the SciDB's
  coordinator.} }
}
\author{
Alber Sanchez
}

