\docType{methods}
\name{create1DModisArray}
\alias{create1DModisArray}
\title{Create an array of 1 dimmension in SciDB}
\usage{
create1DModisArray(object, arrayName)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{arrayName}{Name of the array to be created}
}
\description{
Create an array of 1 dimmension in SciDB
}

