\docType{methods}
\name{create3DModisArray}
\alias{create3DModisArray}
\title{Create an array of 3 dimmension in SciDB}
\usage{
create3DModisArray(object, arrayName)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{arrayName}{Name of the array to be created}
}
\description{
Create an array of 3 dimmension in SciDB
}

