\docType{methods}
\name{createModisArrays}
\alias{createModisArrays}
\title{Creates the arrays needed to store MODIS images}
\usage{
createModisArrays(object, force)
}
\arguments{
  \item{object}{A ScidnInstance object}

  \item{force}{Deletes the arrays (if they exist) and
  create new ones}
}
\description{
Creates the arrays needed to store MODIS images
}

