\docType{methods}
\name{date2grid}
\alias{date2grid}
\title{Converts a date into grid time index}
\usage{
date2grid(object, dateAsText)
}
\arguments{
  \item{object}{Object of the class ModisGrid}

  \item{dateAsText}{Date as a text string}
}
\value{
A date object (POSIXlt)
}
\description{
Converts a date into grid time index
}

