\name{date2ydoy}
\alias{date2ydoy}
\title{Transforms a date into  the year-day_of_the_year date (YYYYDOY)}
\usage{
date2ydoy(object, dateAsText)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{dateAsText}{Date as a text string}
}
\value{
Character representing a date as day-of-the-year (YYYYDOY)
}
\description{
Transforms a date into the year-day_of_the_year date
(YYYYDOY)
}

