\docType{methods}
\name{downloadHdfs}
\alias{downloadHdfs}
\alias{downloadHdfs,object-method}
\title{Downloads MODIS' hdf files}
\value{
  A character vector with the paths to the downloaded files

  A character vector with the paths to the downloaded files
}
\description{
  Downloads MODIS' hdf files

  Downloads MODIS' hdf files
}

