\docType{methods}
\name{downloadHdfs}
\alias{downloadHdfs}
\title{Downloads MODIS' hdf files. Wrapper of the MODIS::getHdf}
\usage{
downloadHdfs(object)
}
\arguments{
  \item{object}{A ModisDownloader object}
}
\value{
A character vector with the paths to the downloaded files
}
\description{
Downloads MODIS' hdf files. Wrapper of the MODIS::getHdf
}

