\docType{methods}
\name{exist}
\alias{exist}
\title{Checks if an array with the given name exists in SciDB}
\usage{
exist(object, arrayName)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{arrayName}{Vector with the names of the arrays}
}
\value{
Logical vector with TRUE if the array exists, FALSE
otherwise
}
\description{
Checks if an array with the given name exists in SciDB
}

