\docType{methods}
\name{findInvalidHdfs}
\alias{findInvalidHdfs}
\title{Finds invalid HDF files}
\usage{
findInvalidHdfs(object, hdfFiles)
}
\arguments{
  \item{object}{A ModisDownloader object}

  \item{hdfFiles}{Character vector of HDF file paths}
}
\value{
A list with the paths of invalid files
}
\description{
Finds invalid HDF files
}

