\name{getFileExtension}
\alias{getFileExtension}
\title{Find the extension of a file}
\usage{
getFileExtension(object, fileName)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{fileName}{Name of the file}
}
\value{
The file extension
}
\description{
Find the extension of a file
}

