\name{getFilenameFromFilepath}
\alias{getFilenameFromFilepath}
\title{Returns the filename of the path to the file}
\usage{
getFilenameFromFilepath(object, filepath)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{filepath}{Character representing the full path to
  the file}
}
\value{
Character representing the filename including the file
extension
}
\description{
Returns the filename of the path to the file
}

