\name{getFileNoExtension}
\alias{getFileNoExtension}
\title{Returns the filename without extension}
\usage{
getFileNoExtension(object, filename)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{filename}{Character representing the file name
  including extension}
}
\value{
Character representing the filename without the file
extension
}
\description{
Returns the filename without extension
}

