\name{getFilepathFromFilepath}
\alias{getFilepathFromFilepath}
\title{Returns the filepath of the path witout the last part (filename)}
\usage{
getFilepathFromFilepath(object, filepath)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{filepath}{Character representing the full path to
  the file}
}
\value{
Character representing the filepath without the file name
}
\description{
Returns the filepath of the path witout the last part
(filename)
}

