\name{getFileresultFromFilename}
\alias{getFileresultFromFilename}
\title{Builds a file name for storing the results of processing}
\usage{
getFileresultFromFilename(object, fileName, band, ext)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{fileName}{Character representing the file name}

  \item{band}{Character representing the band}

  \item{ext}{Character representing the extension for the
  results file (i.e ".txt")}
}
\value{
Character representing a filename
}
\description{
Builds a file name for storing the results of processing
}

