\docType{methods}
\name{getHost}
\alias{getHost}
\title{Returns the object's host property}
\usage{
getHost(object)

getHost(object)
}
\arguments{
  \item{object}{A ModisControl object}

  \item{object}{A ScidbInstance object}
}
\description{
Returns the object's host property

Returns the object host
}

