\name{getTileIdFromFilename}
\alias{getTileIdFromFilename}
\title{Get ths MODIS tile id from the modis filename}
\usage{
getTileIdFromFilename(object, fileName)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{fileName}{Name of the file}
}
\description{
Get ths MODIS tile id from the modis filename
}

