\name{getTimeFromHdfFilename}
\alias{getTimeFromHdfFilename}
\title{Gets the adquisition time of a MODIS HDF file name}
\usage{
getTimeFromHdfFilename(object, hdfFilename)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{hdfFilename}{HDF filename}
}
\value{
Character. A date in the format year and day of the year
YYYYDOY
}
\description{
Gets the adquisition time of a MODIS HDF file name
}

