\docType{methods}
\name{getValidArrayName}
\alias{getValidArrayName}
\title{Returns a valid SciDB's array name}
\usage{
getValidArrayName(object, arrayName)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{arrayName}{Name of the array to be validated}
}
\value{
Character. A valid name
}
\description{
Returns a valid SciDB's array name
}

