\docType{methods}
\name{insert}
\alias{insert}
\title{Insert array data into another array}
\usage{
insert(object, originArray, destinationArray)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{originArray}{Array from where the data is taken}

  \item{destinationArray}{Array where to insert the data}
}
\description{
Insert array data into another array
}

