\docType{methods}
\name{isUploaded}
\alias{isUploaded}
\title{Checks if the given MODIS files has been uploaded}
\usage{
isUploaded(object, modisFiles)
}
\arguments{
  \item{object}{A ModisLoader object}

  \item{modisFiles}{A vector of MODIS HDF filenames}
}
\value{
A character vector
}
\description{
Checks if the given MODIS files has been uploaded
}

