\docType{methods}
\name{listArrays}
\alias{listArrays}
\title{Returns the names of the currrent arrays in the SciDB instance}
\usage{
listArrays(object)
}
\arguments{
  \item{object}{A ScidbInstance object}
}
\value{
A character vecot rof the arrays in SciDB
}
\description{
Returns the names of the currrent arrays in the SciDB
instance
}

