\docType{methods}
\name{loadData}
\alias{loadData}
\title{Loads the input files to SciDB and it deletes the source files}
\usage{
loadData(object)
}
\arguments{
  \item{object}{A Loader object}
}
\value{
A character vector
}
\description{
Loads the input files to SciDB and it deletes the source
files
}

