\docType{methods}
\name{loadDataLong}
\alias{loadDataLong}
\title{Loads the input files to SciDB and it deletes the source files. It doesn't load in parallel}
\usage{
loadDataLong(object)
}
\arguments{
  \item{object}{A Loader object}
}
\value{
A character vector
}
\description{
Loads the input files to SciDB and it deletes the source
files. It doesn't load in parallel
}

