\docType{class}
\name{ModisControl}
\alias{ModisControl}
\alias{ModisControl-class}
\title{The MODISCONTROL class}
\description{
Use this class for reviwing what has been loaded and
processed
}
\details{
You can ???????
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{host}:}{Object of class
  \code{"character"}, it is a parameter for connecting to a
  Database.} \item{\code{dbname}:}{Object of class
  \code{"character"}, it is a parameter for connecting to a
  Database.} \item{\code{user}:}{Object of class
  \code{"character"}, it is a parameter for connecting to a
  Database.} \item{\code{password}:}{Object of class
  \code{"character"}, it is a parameter for connecting to a
  Database.} }
}
\author{
Alber Sanchez
}

