\docType{methods}
\name{getTimeWindowStart}
\alias{getCollections}
\alias{getRequestedProducts}
\alias{getRequestedTiles}
\alias{getTimeWindowEnd}
\alias{getTimeWindowStart}
\title{Returns the object's start time}
\usage{
getTimeWindowStart(object)

getTimeWindowEnd(object)

getRequestedTiles(object)

getRequestedProducts(object)

getCollections(object)
}
\arguments{
  \item{object}{A ModisDownloader object}

  \item{object}{A ModisDownloader object}

  \item{object}{A ModisDownloader object}

  \item{object}{A ModisDownloader object}

  \item{object}{A ModisDownloader object}
}
\description{
Returns the object's start time

Returns the object's end time

Returns the object's requested tiles

Returns the object's requested products

Returns the object's requested collections
}

