\docType{class}
\name{ModisLoader}
\alias{ModisLoader}
\alias{ModisLoader-class}
\title{The MODISLOADER class}
\description{
Use this class for loading to SciDB processed MODIS tiles
}
\details{
You can ???????
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{files}:}{Object of class
  \code{"character"}, it is a vector with the paths to the
  files.} }
}
\author{
Alber Sanchez
}

