\docType{methods}
\name{getFiles}
\alias{getFiles}
\alias{getResultFolder}
\title{Returns the object files}
\usage{
getFiles(object)

getResultFolder(object)
}
\arguments{
  \item{object}{A ModisProcessor object}

  \item{object}{A ModisProcessor object}
}
\description{
Returns the object files

Returns the object's result folder path
}

