\docType{class}
\name{ModisProcessor}
\alias{ModisProcessor}
\alias{ModisProcessor-class}
\title{The MODISPROCESSOR class}
\description{
Use this class for processing downloaded MODIS tiles in
order to produce files formated for uploading to SciDB.
}
\details{
You can ???????
}
\note{
No notes
}
\section{Slots}{
  \describe{ \item{\code{files}:}{Object of class
  \code{"character"}, it is a vector with the paths to the
  HDF files.} \item{\code{resultFolder}:}{Object of class
  \code{"character"}, it is tghe path to the folder for
  storing results.}#' }
}
\author{
Alber Sanchez
}

