\docType{class}
\name{Processor}
\alias{Processor}
\alias{Processor-class}
\title{The PROCESSOR class}
\description{
Use this class for processing data files in order to
produce files formated for SciDB.
}
\note{
No notes
}
\section{Slots}{
  \describe{ \item{\code{files}:}{Object of class
  \code{"character"}, it is a vector with the paths to the
  input files.} \item{\code{resultFolder}:}{Object of class
  \code{"character"}, it is the path to the folder for
  storing the resulting files.}' }
}
\author{
Alber Sanchez
}

