\docType{methods}
\name{queryAqlCmd}
\alias{queryAqlCmd}
\title{Sends an AQL query to SciDB using the OS instead of scidbR (Only works for local installations)}
\usage{
queryAqlCmd(object, aql)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{aql}{AQL expression}
}
\description{
Sends an AQL query to SciDB using the OS instead of scidbR
(Only works for local installations)
}

