\name{substrRight}
\alias{substrRight}
\title{Get the characters of a string from right to left - taken from http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r}
\usage{
substrRight(object, x, n)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{x}{A string}

  \item{n}{The number of chars}
}
\value{
The requested characters
}
\description{
Get the characters of a string from right to left - taken
from
http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r
}

