\name{text2date}
\alias{text2date}
\title{Transforms a date given as text to a date object}
\usage{
text2date(object, dateAsText)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{dateAsText}{Date as a text string}
}
\value{
A date object (POSIXlt)
}
\description{
Transforms a date given as text to a date object
}

