\name{ydoy2date}
\alias{ydoy2date}
\title{Transforms a date in the year-day_of_the_year format to a date}
\usage{
ydoy2date(object, YYYYDOY)
}
\arguments{
  \item{object}{An instance of the class Util}

  \item{YYYYDOY}{Character with 4 digits for the year and 3
  for the day of the year (i.e 2012324)}
}
\value{
A date object
}
\description{
Transforms a date in the year-day_of_the_year format to a
date
}

