\docType{methods}
\name{insert}
\alias{insert}
\title{Insert one array into other}
\usage{
insert(object, originArray, destinationArray)
}
\arguments{
  \item{object}{A ScidbInstance object}

  \item{originArray}{Name of the source array}

  \item{destinationArray}{Name of the target array}
}
\description{
Insert one array into other
}

