\docType{class}
\name{ModisLoader}
\alias{ModisLoader}
\alias{ModisLoader-class}
\title{The MODISLOADER class}
\description{
Use this class for loading to SciDB processed MODIS tiles
}
\details{
You can ???????
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{files}:}{Object of class
  \code{"character"}, it is a vector with the paths to the
  files.} \item{\code{scidbhost}:}{Object of class
  \code{"character"}, it is the name of the host of a SciDB
  instance.} \item{\code{scidbport}:}{Object of class
  \code{"character"}, it is the number of the portof a
  SciDB instance.} \item{\code{scidbInstance}:}{Object of
  class \code{"ScidbInstance"}, it holds the connection and
  related operations to a SciDB instance.} }
}
\author{
Alber Sanchez
}

