\name{ModisProcessor}
\alias{ModisProcessor}
\alias{ModisProcessor-class}
\title{The MODISPROCESSOR class}
\description{
  Use this class for processing downloaded MODIS tiles in
  order to produce files formated for uploading to SciDB.
}
\details{
  You can ???????
}
\note{
  No notes
}
\section{Slots}{
  \describe{ \item{\code{localArcPath}:}{Object of class
  \code{"character"}, it is the path to the folder where
  the MODIS files are stored. It is the path to the
  "MODIS_ARC" folder. See the MODIS package documentation -
  MODISoptions} }
}
\author{
  Alber Sanchez
}

