\docType{class}
\name{ScidbInstance}
\alias{ScidbInstance}
\alias{ScidbInstance-class}
\title{The SCIDBINSTANCE class}
\description{
Use this class for
}
\details{
You can ???????
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{files}:}{Object of class
  \code{"character"}, it is a vector with the paths to the
  files.} }
}
\author{
Alber Sanchez
}

