\docType{class}
\name{ScidbInstance}
\alias{ScidbInstance}
\alias{ScidbInstance-class}
\title{The SCIDBINSTANCE class}
\description{
Use this class for doing stuff related to SciDB
}
\note{
No notes
}
\section{Slots }{
  \describe{ \item{\code{host}:}{Object of class
  \code{"character"}, it is the name of the host of a SciDB
  instance.} \item{\code{port}:}{Object of class
  \code{"character"}, it is the number of the portof a
  SciDB instance.} }
}
\author{
Alber Sanchez
}

